requirements.txt
grequests
pyfiglet
requests
elasticsearch
python-logstash-async
python-dotenv